Port registers to ATtiny85, & get to compile
NB: Oops! I just realized that we ported the buzzer to Timer 0, which
is already used (I think) by `millis()`. So, we need to move it to
Timer 1 instead! Keep in mind Timer 1 is a 16-bit timer so you'll need
to account for that, possibly even downgrading it to act like an 8-bit
timer, though I'd have to look into that more still to be sure.